RnD/requirements.txt
asgiref==3.8.1
brisque==0.0.16
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
fonttools==4.58.4
image==1.5.33
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
libsvm-official==3.30.0
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
tifffile==2025.6.11
tzdata==2025.2